refactor(normalization): address code-review nitpicks and document shared guards/math flow
- replace magic chained-operator pass limit in `MathExpressionNormalizer` with named constant (`maxChainedOperatorPasses`)
- extract duplicated code-ish line detection into shared `CodeishLineDetector`
- remove duplicated `assignmentLikeRegex` and inline code-ish logic from:
- `SentenceCapitalizationNormalizer`
- `MathExpressionNormalizer`
- delegate both normalizers to the shared detector to keep behavior aligned
- wire `CodeishLineDetector.swift` into Xcode project sources
- update docs to reflect new normalization components and pipeline order:
- `Docs/CODEMAP.md` adds `MathExpressionNormalizer` and `CodeishLineDetector` entries
- `Docs/ENGINEERING.md` updates normalization architecture summary and post-processing order (`colon -> math -> list -> ...`)
