+++ Core/Normalization/SentenceCapitalizationNormalizer.swift
@@ -21,10 +21,6 @@ struct SentenceCapitalizationNormalizer {
         pattern: #"(?<![A-Za-z0-9_])i(?![A-Za-z0-9_])"#,
         options: []
     )
-    private static let assignmentLikeRegex: NSRegularExpression? = try? NSRegularExpression(
-        pattern: #"\b[A-Za-z_][A-Za-z0-9_]*\s*=\s*"#,
-        options: []
-    )
     private static let commonTopLevelDomains: Set<String> = [
         "com", "net", "org", "io", "app", "dev", "ai", "co", "me", "edu", "gov",
         "us", "uk", "ca", "au", "de", "fr", "jp"
@@ -254,30 +250,7 @@ struct SentenceCapitalizationNormalizer {
     }
 
     private func isLikelyCodeishLine(_ line: String) -> Bool {
-        guard !line.isEmpty else { return false }
-        if line.contains("`") { return true }
-
-        let codeOperators = ["==", "!=", "<=", ">=", "=>", "->", "::", "&&", "||", "++", "--"]
-        if codeOperators.contains(where: line.contains) {
-            return true
-        }
-
-        if let assignmentRegex = Self.assignmentLikeRegex {
-            let nsLine = line as NSString
-            let range = NSRange(location: 0, length: nsLine.length)
-            if assignmentRegex.firstMatch(in: line, options: [], range: range) != nil {
-                return true
-            }
-        }
-
-        if line.contains(";") {
-            let hasBracesOrBrackets = line.contains("{") || line.contains("}") || line.contains("[") || line.contains("]")
-            if hasBracesOrBrackets {
-                return true
-            }
-        }
-
-        return false
+        CodeishLineDetector.isLikelyCodeishLine(line)
     }
 
     private func isLikelyDomainBoundary(_ text: String, dotLocation: Int) -> Bool {
